Improved CLI rendering

+++ Chess.Lib/UI/RectInt.cs
@@ -1,10 +1,10 @@
 ﻿namespace Chess.Lib.UI;
 
-public readonly record struct RectInt((int X, int Y) LowerRight, (int X, int Y) UpperLeft)
+public readonly record struct RectInt((long X, long Y) LowerRight, (long X, long Y) UpperLeft)
 {
-    public int Width => Math.Abs(LowerRight.X - UpperLeft.X);
+    public long Width => Math.Abs(LowerRight.X - UpperLeft.X);
 
-    public int Height => Math.Abs(LowerRight.Y - UpperLeft.Y);
+    public long Height => Math.Abs(LowerRight.Y - UpperLeft.Y);
 
     public readonly bool OverlapsWith(in RectInt other)
         => other.LowerRight.X >= UpperLeft.X && other.LowerRight.Y >= UpperLeft.Y && other.UpperLeft.X <= LowerRight.X && other.UpperLeft.Y <= LowerRight.Y;
